Move catalog validation logic to a dedicated CatalogValidator class
