+++ agent_sdks/python/a2ui_core/src/a2ui/core/validating/__init__.py
@@ -19,6 +19,11 @@ from .integrity_checker import (
     validate_recursion_and_paths,
     get_component_references,
 )
+from .catalog_validator import (
+    CatalogValidator,
+    ModelCatalogValidator,
+    JsonCatalogValidator,
+)
 
 __all__ = [
     "A2uiValidator",
@@ -26,4 +31,7 @@ __all__ = [
     "validate_component_integrity",
     "validate_recursion_and_paths",
     "get_component_references",
+    "CatalogValidator",
+    "ModelCatalogValidator",
+    "JsonCatalogValidator",
 ]
